requirements.txt
asyncio==3.4.3
bcrypt==3.2.0
fastapi==0.70.0
httptools==0.3.0
orjson==3.6.4
pyjwt==1.7.1
requests==2.22.0
uvicorn==0.15.0
websockets==10.0